moved yaml stification forwarding into back into yam_file.py
